# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Peexoo's API swagger documentation",
        "description": "This is a sample Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/findmyface/": {
            "put": {
                "tags": [
                    "findmyface"
                ],
                "summary": "Downloads the album to the user's FindMyFace Album Collection",
                "operationId": "download",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            },
            "post": {
                "tags": [
                    "findmyface"
                ],
                "summary": "Searches for the face of a user",
                "operationId": "FindMyFace",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "image to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/findmyface/train": {
            "post": {
                "tags": [
                    "findmyface"
                ],
                "summary": "Trains uploaded images for recognition",
                "operationId": "indexFindMyFace",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "image to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/images/random/{count}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Fetches {count} Random Images",
                "operationId": "getRandomImages",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "Maximum number of random images to fetch",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/images/explore/{count}": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Fetches {count} Random Images",
                "operationId": "getRandomImages",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "Maximum number of random images to fetch",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key_scheme": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}